PLA-7156 move PEW executions
