+++ src/citrine/resources/predictor_evaluation_execution.py
@@ -1,109 +1,20 @@
 """Resources that represent both individual and collections of predictor evaluation executions."""
-from functools import lru_cache, partial
+from functools import partial
 from typing import Optional, Union, Iterator
 from uuid import UUID
+import sys
 
 from citrine._rest.collection import Collection
-from citrine._rest.resource import Resource
-from citrine._rest.asynchronous_object import AsynchronousObject
-from citrine._serialization import properties
 from citrine._session import Session
 from citrine.informatics.modules import ModuleRef
-from citrine.informatics.predictor_evaluation_result import PredictorEvaluationResult
 from citrine.resources.response import Response
+import citrine.informatics.executions
+from citrine.informatics.executions import PredictorEvaluationExecution
+
+from citrine._utils.functions import shadow_classes_in_module
 
 
-class PredictorEvaluationExecution(Resource['PredictorEvaluationExecution'], AsynchronousObject):
-    """The execution of a PredictorEvaluationWorkflow.
-
-    Possible statuses are INPROGRESS, SUCCEEDED, and FAILED.
-    Predictor evaluation executions also have a ``status_description`` field with more information.
-
-    """
-
-    # This should really be _session, but _fetch_page assumes there is a 'pageable' parameter
-    _session: Optional[Session] = None
-    """:str: Unique identifier of the project that contains the workflow execution"""
-    project_id: Optional[UUID] = None
-    """:Optional[UUID]: Unique ID of the project that contains this execution."""
-
-    uid: UUID = properties.UUID('id', serializable=False)
-    """:UUID: Unique identifier of the workflow execution"""
-    evaluator_names = properties.List(properties.String, "evaluator_names", serializable=False)
-    """:List[str]: names of the predictor evaluators that were executed. These are used
-    when calling the ``results()`` method."""
-    workflow_id = properties.UUID('workflow_id', serializable=False)
-    """:UUID: Unique identifier of the workflow that was executed"""
-    predictor_id = properties.UUID('predictor_id', serializable=False)
-    status = properties.Optional(properties.String(), 'status', serializable=False)
-    """:Optional[str]: short description of the execution's status"""
-    status_info = properties.Optional(
-        properties.List(properties.String()),
-        'status_info',
-        serializable=False
-    )
-    """:Optional[List[str]]: human-readable explanations of the status"""
-    experimental = properties.Boolean("experimental", serializable=False, default=True)
-    """:bool: whether the execution is experimental (newer, less well-tested functionality)"""
-    experimental_reasons = properties.Optional(
-        properties.List(properties.String()),
-        'experimental_reasons',
-        serializable=False
-    )
-    """:Optional[List[str]]: human-readable reasons why the execution is experimental"""
-
-    def __init__(self):
-        """Predictor evaluation executions are not directly instantiated by the user."""
-        pass  # pragma: no cover
-
-    def __str__(self):
-        return '<PredictorEvaluationExecution {!r}>'.format(str(self.uid))
-
-    def _path(self):
-        return '/projects/{project_id}/predictor-evaluation-executions/{execution_id}' \
-            .format(project_id=self.project_id,
-                    execution_id=self.uid)
-
-    def in_progress(self) -> bool:
-        """Whether predictor evaluation execution is in progress. Does not query state."""
-        return self.status == "INPROGRESS"
-
-    def succeeded(self) -> bool:
-        """Whether predictor evaluation execution has completed successfully. Does not query state."""  # noqa: E501
-        return self.status == "SUCCEEDED"
-
-    def failed(self) -> bool:
-        """Whether predictor evaluation execution has completed unsuccessfully. Does not query state."""  # noqa: E501
-        return self.status == "FAILED"
-
-    @lru_cache()
-    def results(self, evaluator_name: str) -> PredictorEvaluationResult:
-        """
-        Get a specific evaluation result by the name of the evaluator that produced it.
-
-        Parameters
-        ----------
-        evaluator_name: str
-            Name of the evaluator for which to get the results
-
-        Returns
-        -------
-        PredictorEvaluationResult
-            The evaluation result from the evaluator with the given name
-
-        """
-        params = {"evaluator_name": evaluator_name}
-        resource = self._session.get_resource(self._path() + "/results", params=params)
-        return PredictorEvaluationResult.build(resource)
-
-    def __getitem__(self, item):
-        if isinstance(item, str):
-            return self.results(item)
-        else:
-            raise TypeError("Results are accessed by string names")
-
-    def __iter__(self):
-        return iter(self.evaluator_names)
+shadow_classes_in_module(citrine.informatics.executions, sys.modules[__name__])
 
 
 class PredictorEvaluationExecutionCollection(Collection["PredictorEvaluationExecution"]):
